ci: migrate benchmark tool benny to vitest
