+++ package.json
@@ -14,14 +14,12 @@
     "@typescript-eslint/eslint-plugin": "^7.16.1",
     "@typescript-eslint/parser": "^7.16.1",
     "@vitest/coverage-istanbul": "^2.0.5",
-    "benny": "^3.7.1",
     "es-toolkit": "^1.26.1",
     "eslint": "^9.7.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-jsdoc": "^48.7.0",
     "globals": "^15.11.0",
     "prettier": "^3.3.3",
-    "ts-node": "^10.9.2",
     "ts-pattern": "^5.5.0",
     "tsup": "^8.2.3",
     "typescript": "^5.5.3",
@@ -30,11 +28,9 @@
   "scripts": {
     "build": "tsup ./src/index.ts --format cjs,esm --dts --minify",
     "test": "vitest run --coverage --typecheck",
+    "bench": "vitest bench",
     "lint": "eslint \"src/**/*.ts\" --fix",
-    "format": "prettier --write .",
-    "benchmark": "ts-node ./benchmarks/always-last-digit.ts && ts-node ./benchmarks/random-digit.ts",
-    "bench-always-last-digit": "ts-node ./benchmarks/always-last-digit.ts",
-    "bench-random-digit": "ts-node ./benchmarks/random-digit.ts"
+    "format": "prettier --write ."
   },
   "main": "./dist/index.js",
   "browser": "./dist/browser.global.js",
