+++ tsconfig.json
@@ -4,7 +4,7 @@
      * ------------------------------------------------------------------------------------------------------------------------------------------------ */
     // "incremental": true,                   /* 증분 컴파일 활성화 */
     "target": "ESNext" /* ECMAScript 목표 버전 설정: 'ES3'(기본), 'ES5', 'ES2015', 'ES2016', 'ES2017','ES2018', 'ES2019', 'ES2020', or 'ESNEXT'. */,
-    "module": "CommonJS" /* 생성될 모듈 코드 설정: 'none', 'commonjs', 'amd', 'system', 'umd', 'es2015', 'es2020', or 'ESNext'. */,
+    "module": "ESNext" /* 생성될 모듈 코드 설정: 'none', 'commonjs', 'amd', 'system', 'umd', 'es2015', 'es2020', or 'ESNext'. */,
     "lib": ["ESNext"] /* 컴파일 과정에 사용될 라이브러리 파일 설정 */,
     //  "allowJs": true,                          /* JavaScript 파일 컴파일 허용 */
     //  "checkJs": true,                       /* .js 파일 오류 리포트 설정 */
